+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ------------------------------------------------------------------------
 - Fix: [#1281] The construction window won't open under some circumstances.
 - Fix: [#1942] Crash when attempting to clear a building with the clear construction tool.
+- Fix: [#1946] Crash when loading openloco.yml config that has no shortcuts defined.
 - Technical: [#1934] The performance of scrollable lists has been improved.
 
 23.04.1 (2023-04-27)